+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - [feat: support using seperated cb in wildcard-api level and add cb counters expire interval config](https://github.com/Tencent/spring-cloud-tencent/pull/1809)
 - [feat: adapt to polaris-java ReportClientRequestCustomizer plugin for config watch reporting](https://github.com/Tencent/spring-cloud-tencent/pull/1813)
 - [feat: support audit log](https://github.com/Tencent/spring-cloud-tencent/pull/1812)
+- [feat: support config effective value ](https://github.com/Tencent/spring-cloud-tencent/pull/1815)